Set profile=server as node default
In hope this is the right way to do it, take 2? From my experience, by having this as node default it will benefit all members. VPS providers does not like local lan scanning, why waste time in configuring it? In-home routers struggle with local lan scanning up to the point they are failing. It will be easier for those who need it (2% ?), to turn scanning on, than 98% to turn it off.